maps with assent
I use some of the textures that was ready to use in UE4 , I duplicated some of them and change their colors in material_boxes folder.
I used some of the shapes that are in shape folder.
The traps I use are saved in Spike folder.
I created an elevator and placed it in rogue_lite folder.
The material folder is the grass and rockes that I used to paint the Landscape.
There are 3 static meshes in rogue_lite folder that I created for level s stairs .